requirements.txt
BeautifulSoup==3.2.1
PySocks==1.5.0
argparse==1.2.1
beautifulsoup4==4.3.2
wsgiref==0.1.2
